+++ src/cow-react/modules/application/pure/Widget/index.cosmos.tsx
