+++ src/cow-react/pages/Account/Tokens/styled.ts
@@ -1,6 +1,6 @@
 import { ChevronDown } from 'react-feather'
 import styled from 'styled-components/macro'
-import { Content } from 'cow-react/modules/application/dumb/Page'
+import { Content } from 'cow-react/modules/application/pure/Page'
 import { ThemedText, MEDIA_WIDTHS } from 'theme'
 import { Card } from 'cow-react/pages/Account/styled'
 import { darken, transparentize } from 'polished'
