+++ src/cow-react/pages/Account/index.tsx
@@ -2,7 +2,7 @@ import { lazy, Suspense } from 'react'
 import { Switch, Route, useLocation } from 'react-router-dom'
 import { AccountMenu } from './Menu'
 import { Wrapper, AccountPageWrapper } from './Tokens/styled'
-import { Content, Title } from 'cow-react/modules/application/dumb/Page'
+import { Content, Title } from 'cow-react/modules/application/pure/Page'
 import { Routes } from 'cow-react/constants/routes'
 import { Loading } from 'components/FlashingLoading'
 import { Container, CardsWrapper } from './styled'
