+++ src/cow-react/pages/CookiePolicy/index.tsx
@@ -1,6 +1,6 @@
 import contentFile from './CookiePolicy.md'
 import { MarkdownPage } from 'components/Markdown'
-import { GdocsListStyle } from 'cow-react/modules/application/dumb/Page'
+import { GdocsListStyle } from 'cow-react/modules/application/pure/Page'
 import styled from 'styled-components/macro'
 
 const Wrapper = styled(MarkdownPage)`
