+++ src/cow-react/pages/Faq/ToC.tsx
@@ -1,4 +1,4 @@
-import { Content, Title } from 'cow-react/modules/application/dumb/Page'
+import { Content, Title } from 'cow-react/modules/application/pure/Page'
 import { LinkScrollable } from 'components/Link'
 import { TocSection } from '.'
 import { PageIndex } from './styled'
