+++ src/cow-react/pages/Faq/TokenFaq.tsx
@@ -1,4 +1,4 @@
-import { Page, Content } from 'cow-react/modules/application/dumb/Page'
+import { Page, Content } from 'cow-react/modules/application/pure/Page'
 
 import { ExternalLinkFaq, Wrapper } from './styled'
 import { Link } from 'react-router-dom'
