+++ src/cow-react/pages/error/AnySwapAffectedUsers/index.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components/macro'
-import { Page, Title, Content, GdocsListStyle } from 'cow-react/modules/application/dumb/Page'
+import { Page, Title, Content, GdocsListStyle } from 'cow-react/modules/application/pure/Page'
 import { ButtonPrimary } from 'custom/components/Button'
 import cow404IMG from 'assets/cow-swap/cow-404.png'
 import { ExternalLink as ExternalLinkTheme } from 'theme'
