+++ src/cow-react/pages/error/NotFound/index.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom'
 import styled from 'styled-components/macro'
-import { Page, Title, Content, GdocsListStyle } from 'cow-react/modules/application/dumb/Page'
+import { Page, Title, Content, GdocsListStyle } from 'cow-react/modules/application/pure/Page'
 import { ButtonPrimary } from 'custom/components/Button'
 import cow404IMG from 'assets/cow-swap/cow-404.png'
