+++ src/cow-react/pages/games/MevSlicer/index.tsx
@@ -1,5 +1,5 @@
 import { ButtonPrimary } from 'components/Button/ButtonMod'
-import { Page, Content } from 'cow-react/modules/application/dumb/Page'
+import { Page, Content } from 'cow-react/modules/application/pure/Page'
 import styled from 'styled-components/macro'
 
 import ninjaCowImg from 'assets/cow-swap/ninja-cow.png'
