+++ src/custom/components/ErrorBoundary/ErrorWithStackTrace.tsx
@@ -7,7 +7,7 @@ import { ExternalLink } from 'theme'
 import { CODE_LINK, DISCORD_LINK } from 'constants/index'
 import React from 'react'
 import styled from 'styled-components/macro'
-import { Title } from 'cow-react/modules/application/dumb/Page'
+import { Title } from 'cow-react/modules/application/pure/Page'
 import { userAgent } from '@src/utils/userAgent'
 import store, { AppState } from 'state'
