+++ src/custom/components/Markdown/index.tsx
@@ -3,7 +3,7 @@ import ReactMarkdownHtml from 'react-markdown/with-html'
 import ReactMarkdown, { ReactMarkdownPropsBase } from 'react-markdown'
 import useFetchFile from 'hooks/useFetchFile'
 import { HeadingRenderer } from './renderers'
-import { Page, Title, Content } from 'cow-react/modules/application/dumb/Page'
+import { Page, Title, Content } from 'cow-react/modules/application/pure/Page'
 import styled from 'styled-components/macro'
 import { WithClassName } from 'types'
 import { LinkScrollable, Link } from 'components/Link'
